feat: modularize frontend scripts
